refactor(health): per-section components + independent fetching for the health page
The 1,137-line HealthPage — one component whose ~9 fetches gated the
whole page behind a single Promise.all + loading flag, blanking every
panel on one fault — becomes a ~110-line layout shell over per-section
components that each fetch independently:

- use-health-data.ts — usePolledJson<T>, a per-section polled JSON GET
  with its own { data, error, loading, lastUpdated, refresh } lifecycle
  (optional-endpoint swallow + a select validator for reject-and-keep).
- health-sections.tsx — SummaryCards, SpendTokenSection, UsageSection,
  ContextUsageSection, DiagnosticsSection, ServerFooter + the shared leaf
  helpers (HorizontalBars, ListSearch, usage panels, STATUS_STYLES).
- search-section.tsx — self-contained: own search-status/telemetry poll,
  the blue/green reindex convergence loop, and the SSE-driven
  usePluginEvent reindex-progress wiring.
- plugins-section.tsx — self-contained: registry + manifest poll, the
  filter, and the upgrade modal with its consent flow.

DELIBERATE behavior change: loading is no longer all-or-nothing. Each
section renders its own skeleton/empty/hidden state, and a single failing
endpoint faults only its own card while the rest of the page keeps
working — SummaryCards surfaces a per-section error instead of the whole
page showing "Failed to load health data". The header Refresh button now
bumps a shared refreshNonce that re-triggers every section. Summary and
usage feeds are fetched once in the shell and shared by the sections that
consume them (cards/diagnostics/footer; token/context). JSX is otherwise
moved verbatim.

The existing health-page tests pass unchanged — they assert rendered data
(per-table doc counts, the token table, plugin update controls), not the
old single-loading-gate shape, so no assertions were weakened.
